Add offline version of Titan
Use the newly added _get_weights_path method to use an optional
weights_path given during init or look at the path in the model
registry.
If a valid path is found, the load the offline model from there.
Otherwise, ensure that there is an internet access, then use the
online version of Titan.